Split Osmosis management code away from osmose_run.py
